+++ src/main/java/com/powsybl/iidm/converter/server/IidmConverterSwaggerConfig.java
