+++ src/main/java/com/powsybl/iidm/converter/server/dto/NetworkIds.java
