+++ src/test/java/com/powsybl/iidm/converter/server/IidmConverterTest.java
@@ -1,0 +1,84 @@
+package com.powsybl.iidm.converter.server;
+
+import com.powsybl.commons.datasource.ReadOnlyDataSource;
+import com.powsybl.iidm.import_.Importers;
+import com.powsybl.iidm.network.Network;
+import com.powsybl.network.store.client.NetworkStoreService;
+import org.apache.commons.compress.utils.IOUtils;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.web.client.RestTemplate;
+
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.util.UUID;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.ArgumentMatchers.any;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@RunWith(SpringRunner.class)
+@WebMvcTest(IidmConverterController.class)
+@ContextConfiguration(classes = {IidmConverterApplication.class, IidmConverterService.class})
+public class IidmConverterTest {
+
+    @Autowired
+    private MockMvc mvc;
+
+    @Mock
+    private RestTemplate caseServerRest;
+
+    @Mock
+    private RestTemplate geoDataRest;
+
+    @Autowired
+    private IidmConverterService iidmConverterService;
+
+    @MockBean
+    private NetworkStoreService networkStoreClient;
+
+    @Before
+    public void setUp() {
+        MockitoAnnotations.initMocks(this);
+    }
+
+    @Test
+    public void test() throws Exception {
+        ClassLoader classLoader = getClass().getClassLoader();
+        try (InputStream inputStream = classLoader.getResourceAsStream("testCase.xiidm")) {
+            byte[] networkByte = IOUtils.toByteArray(inputStream);
+            iidmConverterService.setCaseServerRest(caseServerRest);
+            iidmConverterService.setGeoDataRest(geoDataRest);
+            given(caseServerRest.exchange(any(String.class), any(HttpMethod.class), any(HttpEntity.class), any(Class.class)))
+                    .willReturn(new ResponseEntity<>(networkByte, HttpStatus.OK));
+            InputStream byteArrayInputStream = new ByteArrayInputStream(networkByte);
+            Network network = Importers.loadNetwork("test.xiidm", byteArrayInputStream);
+            given(networkStoreClient.importNetwork(any(ReadOnlyDataSource.class))).willReturn(network);
+            UUID randomUuid = UUID.fromString("78e13f90-f351-4c2e-a383-2ad08dd5f8fb");
+            given(networkStoreClient.getNetworkUuid(network)).willReturn(randomUuid);
+
+            MvcResult mvcResult = mvc.perform(post("/v1/persistent-store/{caseName}", "testCase.xiidm"))
+                    .andExpect(status().isOk())
+                    .andReturn();
+
+            assertEquals("{\"networkUuid\":\"78e13f90-f351-4c2e-a383-2ad08dd5f8fb\",\"networkId\":\"20140116_0830_2D4_UX1_pst\"}",
+                    mvcResult.getResponse().getContentAsString());
+        }
+    }
+}
